chore: merge windows.js + refactor
